refactor(StringAddCalculator) : StringAddCalculator 및 테스트 클래스의 패키지 변경
