+++ src/test/java/calculator/StringAddCalculatorTest.java
@@ -1,5 +1,6 @@
-package study;
+package calculator;
 
+import calculator.StringAddCalculator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
